Renamed "service" folder to "services"
