+++ test/main/java/services/LocalSearchServiceTest.java
@@ -1,4 +1,4 @@
-package main.java.service;
+package main.java.services;
 
 import main.java.api.DroneApiService;
 import main.java.api.IDroneApiService;
